pyproject.toml
[project]
name = "mlops-lab3"
version = "0.1.0"
description = "MLOps Lab 3 - Model Training with MLFlow and ONNX"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "black>=25.11.0",
    "click>=8.1.8",
    "fastapi>=0.121.3",
    "flake8>=7.0.0",
    "gradio>=4.0.0",
    "httpx>=0.28.1",
    "isort>=5.13.0",
    "jinja2>=3.1.6",
    "ml-dtypes>=0.5.0",
    "mlflow>=2.10.0",
    "mypy>=1.8.0",
    "numpy>=1.26.0,<2.0.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "onnxscript>=0.1.0",
    "pillow>=11.3.0",
    "protobuf>=4.21.0",
    "pylint>=3.3.9",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.9",
    "requests>=2.31.0",
    "uvicorn>=0.38.0",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["api", "cli", "logic"]
known_third_party = ["click", "fastapi", "gradio", "mlflow", "numpy", "onnx", "onnxruntime", "PIL", "pytest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=logic",
    "--cov=cli",
    "--cov=api",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.coverage.run]
source = ["logic", "cli", "api"]
omit = ["*/tests/*", "*/venv/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
